sched/fair.c: Don't allow SchedTune boosted tasks to be migrated to small cores during interactions
We want boosted tasks to run on big cores. But CAF's load balancer
changes do not account for SchedTune boosting, so this allows for
boosted tasks to be migrated to a suboptimal core. Let's mitigate
this by setting the LBF_IGNORE_STUNE_BOOSTED_TASKS for tasks
migrating from a larger capacity core to a min capacity one and
that have a schedtune boost > 10. If both are true, do
not migrate this task. If the next time the load balancer runs,
the same task is selected, we clear the
LBF_IGNORE_STUNE_BOOSTED_TASKS flag.

Signed-off-by: Zachariah Kennedy <zkennedy87@gmail.com>
Signed-off-by: Anirudh Gupta <anirudhgupta109@gmail.com>